Problem: omni's alter extension support availability
If `omni` extension itself is not installed in the database, its
handling of `alter extension` is not going to be invoked. This is
because it uses our own initialization system to register the hooks.

Solution: hard-code loading omni module itself

This is less than perfect, but solves the problem for now. Once we
consider upgradeability of omni itself, things may change and shuffle a
bit.