Problem: only package and release extension versions changed in a PR
It is not trivial to do it by storing extension versions
in CMakeLists.txt.

Solution: Store current extension versions in 'versions.txt' file.
Each line in it will be in the form of <ext-name>=<ext-ver>. Any
extension not specified in it will be assigned a special 'unreleased'
version. This file will be used to compute the extension versions
changed in a PR.

Signed-off-by: Rahul Modpur <rmodpur2@gmail.com>