Problem: does omni expose xact_callback hook yet?
We have the hook iterating over the dynamic hooks, but I am not sure it
owrks.

Solution: ensure we test the case

Importantly, this fixes a segfault in hook registration when no entry
point has been allocated, which may be the case with `xact_callback` as
we don't anything there by default.