Problem: omni_httpc does not follow HTTP redirects
Some APIs (or websites) redirect elsewhere before serving any content.
Implementing a logic that requires making another request in this case
is bothersome, especially if it needs to be done in many places.

Solution: implement `follow_redirects` option

The way it is implemented it is that in the on_head callback (that h2o
calls us when the HTTP response status line and headers have been
read), we check the headers for a location header. Additionally, we
check that the status code is one of 301, 302 or 307, 308. If both of
these hold, we schedule a new HTTP request on the event loop, while
waiting within the current HTTP request on_head to finalize.  This
should allow any other concurrent HTTP requests to run as well, as we
yield the control to the event loop until the redirect requests are
done. This process applies in a recursive fashion, because any
subsequent request might contain another redirect. Once a scheduled
request finishes, we copy the results back to the original request, so
we could send the redirected response back to the user.

An alternative approach would be to not wait in a loop for the request
to finish, but to schedule a new request whilist immediately returning
from the current request and let the callbacks populate the complete
field for the original request. In this case we would need to have
self-referential pointer within the request struct to the original
request or to the previous request. Once a non-redirect request
finishes, it would need to copy the response headers, body, status code
etc. back to the initial request.  However, I think this would make the
code more difficult to understand than the current solution.

To avoid getting stuck in re-direct loops, I added a limit of 5 requests
as suggested by RFC2068
https://datatracker.ietf.org/doc/html/rfc2068#section-10.3 (this is an
older RFC, but I think this simple solution is sufficient in practice).
After reaching this limit no further requests would be attempted. This
is covered by one of the tests.

Closes #444