Problem: clarity on single-shot xact callbacks
Some of the code is not very clear on how single-shot xact callbacks
operate. This is because they were a bit of a hack to begin with and we
might want to improve this part of the extension soon.

But it's still difficult to understand how they operate

Solution: reorganize the code a bit, provide commentary

Also, got rid of unnecessary memory context switching as deletion from a
list uses list's own memory context.