Problem: upgrading omni 0.1.0 to 0.1.1
In CI, testing altering omni from 0.1.0 to 0.1.1 crashes. This is not
good!

Solution: ensure mismatching omni versions can't be loaded

By mismatching versions, I mean version of the preloaded library and the
extension.

This change introduces "module information" which consists of:
descriptive name and versions as well as permanent identity (typically,
a UUID) that allows us to track the "sameness" of the module across
renames. This is used by `omni` to detect if it is attempting to load
`omni`.

As a special case for omni 0.1.0, we're using the presence of
`database_worker`, `startup_worker` and `deinitialize_backend` symbols
as markers as there's no module information there (and we enforce the
fact that there should be no module information in such a module).

Since what we're trying to prevent is loading libraries that
haven't been preloaded (as they need access to the same address space),
we check if the `_Omni_magic` symbols is mapped to the same address as
the host one.

After that, we establish if we have a simple version mismatch as per
module information or whether we're loading a different file.

When omni is mismatches, an error is raised.

Some of this logic has been tested manually, but we're light on
automated testing of this.

There is an important omni upgrade issue/concern: a proper upgrade
procedure is to change the preloaded library, restart Postgres and then
upgrade the extension. This still means that until such upgrade is
performed, the extension is unusable and will error out. This is not
quite perfect.