Problem: "too many dynamic shared memory segments"
This error is happening after some use of omni_httpd.

Solution: use a single DSA

The crux of the problem:

* Number of DSMs is very limited:
  https://github.com/postgres/postgres/blob/master/src/backend/storage/ipc/dsm.c#L195-L196
* We pin DSAs in each backend so that other backends can use memory
  allocated there even if those backends exit
* These numbers run out pretty fast, resulting in the error. A new
  connection, upon being established, creates a new DSA, therefore
  making us come closer to the limit.
* The problem here is that we don’t release these DSA/DSMs ever and
  hit the limits pretty soon as we store pointers

This is unsustainable.

Instead, we now use a single DSA initialized by the first backend and
let other backends attach to it. New DSMs will still be created but at a
much slower rate.