Problem: can't upgrade omni_httpd from 0.1.0 to 0.1.1
The error:

```
could not find function "sum_in" in file "omni_httpd--0.1.1.so"
```

Solution: ensure extension upgrade hook only updates module_pathname

Originally, the hook would look into any pg_proc dependency of the
extension that is implemented in C and would reassign that to a new
module pathname.

However, this doesn't always work well. In the particular example of
omni_httpd, it uses omni_types to create a sum type. It is pointing to
omni_types' library. However, since it is created during omni_http
installation, it is also listed as a dependency.

We now enforce a rule that only procedures that had probin set to the previous
version of the module module_pathname will be changed to the new
module_pathname.

While working on this (relatively simple) change, I ran into unexpected
failures/crashes.

Turns out, hook registration was not implemented correctly. Shifting of the
indexes to prepend an element absolutely did not do what it was supposed
to do. To that end, that portion of hook registration was fixed and some
additional asserts were added for a sanity check.

I've also changed PostgreSQLExtension.cmake a bit to be able to get
dependencies accessed from extensions/omni/omni_test