Problem: Omnigres docker image doesn't pick up current omni version
Omni version is hardcoded in docker-entrypoint-initdb.d sql file
containing `shared_preload_libraries`. On release of new omni
version the container doesn't start because it points to older
omni version which is not present in the container image.

Solution: convert docker-entrypoint-initdb.d scripts to templates
and substitute the env variables used in it using `envsubst`

Signed-off-by: Rahul Modpur <rmodpur2@gmail.com>