Problem: omni_httpd worker sporadically stops listening
This is manifesting in scenarios like recursive HTTP calls (HTTP handler
calling back into the same instance of omni_httpd) and having omni_httpd
not receive an opportunity to accept an incoming connection.

Solution: don't stop the event loop from reading the server socket

This effectively reverses ae72953f406d8716741f0c24f08b0e51321ca20a as it
works better overall in common scenarios.

It doesn't work well when the number of concurrent requests is high
(equal to or higher than omni_httpd.http_workers) and it seemingly gets
overwhelmed while unable to accept the call and it takes time to drain
the queue which results in (effectively a temporary denial of serice)