Problem: omni_python doesn't have a dbapi library
When using omni_python the option to use Python's DB API interface
is not easily available. Users either have to convert their application
to use plpy or use psycopg over the wire.

Solution: Using petere/plpydbapi as a template implement a library in
omni_python that supports this interface. The library is currently
called dbapi.