Problem: deleted omni_vfs local_fs mount
When local_fs information is deleted from `local_fs_mounts`, omni_vfs
would crash. This is not great!

Solution: ensure it checks if there's at least one record
and error out otherwise.

Most likely, the original code was built on assumption that there's
always going to be a value available in this table because it will
always be preceded by `omni_vfs.local_fs()` function call. But that's
actually not true.

Kudos to @rmodpur for spotting this.