Problem: it's unclear how to install omni_httpd
New users run into a case where they install omni_httpd but it doesn't
work.

Solution: provide documentation on install omni as well

It's not great to hardcode versions of omni/omni_httpd in the
documentation, but it's better than no documentation.

In the future, this should be better addressed by omni_httpd warning the
user of the corrective action to be taken.

Better yet, need to continue thinking about simpler installation
procedures for such. Can we ship a version of `omni` with all omni
extensions?