Problem: differences between minor Postgres versions
Most of the time, extensions compiled for one minor version should work
on another, but that's not universally true. There are some changes
between different minor versions that may make the extensions behave
incorrectly or (potentially) cause an abnormality.

Solution: make `omni` module loader issue warnings
when it encounters extensions compiled for a different version of
Postgres.

There's an important implication of this change: we need to change how
we ship the builds of our extensions. We should probably run tests
against all minor revisions, too.