Problem: cmake setup is not suitable for packaging
Currently the cmake setup forces `CPMAddPackage` for various dependencies,
it doesn't allow to get them from a package mananger.

This makes it hard to create a Nix package of omnigres, for one. Other
Linux distributions will have the same problem.

Solution: Try to get dependencies from the system and if not available
use `CPMAddPackage`.

Some dependencies can be found with `find_package` and others with
`pkg_check_modules`.

List of dependencies covered:

- [x] pugixml
- [x] uriparser
- [x] h2o
- [x] libfyaml
- [x] stc
- [x] metalang99