Problem: sporadic failure of redirect test in omni_httpc
Solution: ensure it's clearing the reconfiguration correctly

This is not perfect, as it is obscure and undocumented knowledge, but
this is how it works right now.

The installation of extensions generates two configuration changes, and
we need to clear those before we can ensure the last useful
configuration change is made before we test it.