Problem: managing contextual access to data and functionality
When developing application integrated with the database, it's very
useful to be able to ensure that only permitted data can be accessed,
modified or inserted.

All of this is, of course, possible with Postgres: privileges,
[updateable] views, triggers, RLS.

However, it's difficult to review and maintain policies established by
all these instruments.

Solution: consolidate access policies in a single document

This document can then be imported into a database and the effects of
the policy can be provisioned. This is what the new omni_access
extension does.