Problem: sometimes transactional variables are insufficient
They work well when we handle requests from within Postgres. However,
imagine if one is trying to use RLS from a backend and wants to identify
the user? Holding a transaction may prove less than efficient.

Solution: introduce session-level variables

They behave very similarly to transactional variables but ignore
transactional and sub-transactional boundaries.

This is still not quite perfect as the backend needs to ensure that
connections are not released back to the pool before the operation is
completed and the variables needs to be set to an "invalid" value upon
the return. This will not well at all in statement pooling.