Problem: sharing memory between backends
While we can use DSM/DSA or something similar, it's still painfully
difficult: managing named segments, translating pointers, etc.

Wouldn't it be great to have a shared address space, in anticipation of
Postgres switching to the thread model?

Solution: add the concept of "omni memory" to omni

This omni memory is allocated in the postmaster and is typically equal
to the size of physically available RAM, or `omni.omni_memory_size`
configuration parameter.

On Linux, if configured (`sysctl vm.overcommit_memory` set to 1), we can
significantly overcommit. On macOS, it seems to work out of the box (I
was able to overcommit to the tune of 50TB)

Omni then provides an API to reserve chunks of memory, as well as
iterate over them. Currently, chunk release API is a no-op because
it's a very simple allocation scheme: a bump allocator. It can't free.
This is rather temporary so we don't have to focus on something more
complex.