Problem: access policy joins may duplicate records
When left joining with other tables, if more than one row has been
matched, it'll produce an equivalent number of "duplicates" on the left
side.

Solution: use `distinct` to filter out the duplicates

Is there any better way we can accomplish this? At the very least we can
make it possible to opt out of `distinct` on a join-by-join basis.