Problem: `IsTransactionBlock` check in `omni_txn.retry`
Is it sufficient? I've got a glimpse of it not being effective when
evaluated through SPI. Specifically, it didn't seem to be effective
when a function was called that would call `omni_txn.retry`, even after
doing some modifying statements.

Solution: cover the test cases

So far, they don't seem to confirm the finding – but they are not being
doing through SPI.