Problem: omni_httpd.handlers approach
It limits us as to what would be possible: we always need to execute a
query to serve a request.

So, for example, we won't be able to call a non-atomic procedure, even
if we wanted to.

Solution: switch to direct function call

omni_httpd now exposes a `handler` function that must remain in place
but can be replaced. This function will serve any and all requests.

I am occassionally observing a slightly better RPS with this implementation.