Problem: calling SQL-language omni_httpd.handler
If this function is defined as a SQL function, it will fail with
the violation of assertion for presence of an active snapshot, as they
require these.

Solution: ensure we always have a snapshot to operate in

This requires a kludge for non-atomic functions implemented in languages
like PL/pgSQL if they use SPI. SPI tries to check if portals account
for all active snapshots. But having no portal, they don't, resulting in
this failure:

```
portal snapshots (0) did not account for all active snapshots (1)
```