Problem: uuidv4 is not monotonic
This makes it less acceptable as an identity / primary key as it may
result in poor locality.

Solution: port current patch for UUIDv7 from Postgres CF

https://www.postgresql.org/message-id/flat/CAAhFRxitJv=yoGnXUgeLB_O+M7J2BJAmb5jqAT9gZ3bij3uLDA@mail.gmail.com

As long as it is okay to expose the time of the key creation, UUIDv7 is
a great option for an UUID-based key.

See also:

https://buildkite.com/blog/goodbye-integers-hello-uuids

<blockquote>
With this change (along with several other changes over a 6 week
period), we observed a 50% reduction in the WAL (Write Ahead Log)
rate of the primary database. A similar reduction in write IO was also
observed. This reduction in the WAL rate allowed the team to easily set
up read replicas and perform other database migration tasks.
</blockquote>