Problem: `regproc[edure]` have limited usefulness
Sure, they denote that an OID is a function/procedure, and, in case of
`regprocedure`, that function has certain arguments.

But the type itself does not convey the signature, which makes it not
possible for us to call these function.

Solution: design "function signature types"

This type is also an OID but we also store information about its full
signature and we validate on casts that the functions match the entire
signature.

These makes these types represent the actual function signature, making
them callable. One can compose them into interfaces through composite
types.