Problem: sometimes our current build system fails
And it skips making a release when it happened. The problem is that it
won't go back there anymore and we effectively skip a release.

Solution: designate a manual CI workflow to run against a revision

This way we should be able to get missed releases.

I don't want this to be a long-term solution, but I need to get some
artifacts published!