Problem: checking if function conforms to signature
Currently, we would get an error but that's not great for queries. What
if we want to find all conforming functions?

Solution: introduce TYPE_conforming_function()

This function will return return NULL on non-conforming functions, but
otherwise will work similar to casting to the signature type.

Also, since I was testing on the entire catalog of functions, I made
sure that both conforming_function and casting respect namespaces.