Problem: authenticating users
While not a problem in itself, the fact that every application has to do
it, leads to regular reinvention of the wheel.

Solution: design omni_auth to be a common building block

At the first go, it only supports password-based authentication and only
supports bcrypt for hashing. More methods and hashing algorithms are to
be added later.