Problem: using S3 caching for Docker builds
We've never completed building it out, so there is no automatic
expiration and manual cache cleanup is not very convenient.

Besides, it has left us stranded with an older _forked_ version of
buildkit that was able to read cache from an unauthenticated connection.

Solution: use WarpBuild's caching facility for this instead