Problem: omni_httpd server hangs
This happens when:

* something causes a FATAL error that terminates the background worker
  (such as `transaction_timeout` expiration)
* when SIGTERM is sent to an httpd worker

Solution: ensure Postgres will restart HTTP workers

We removed restarting in e889328f. To the best of my understanding,
since we got rid of dynpgext that was handling restarts, nothing was
restarting these anymore. So if the worker died, they died.

This change also makes sure that HTTP worker exits with a non-zero exit
code when terminated outside of the normal termination sequence.
Adjustments were made both to master worker and http worker to
facilitate the following shutdown scenarios:

1) master worker shutting down

It signals with a special value of MAX_INT in the semaphore that this is
normal termination routine. A bit hacky, but works for now.

2) postmaster being shut down

Here we observe that postmaster amends its _undocumented_ postmaster.pid
to specify that it is being stopped.  We use that to determine that it
is being stopped. Perhaps we should contribute a much simpler method to
signal this condition to the core.
But for the time being...

3) workers being restarted because one of the workers crashed

Here we observe a yet another intricate property: when HTTP workers are
being restarted by postmaster because another worker crashed, they will
no longer have notify_pid (confirmed through reading of
`ResetBackgroundWorkerCrashTimes` code versions 13 to 17). This is
still not perfect as we may (potentially) run into a scenario when we
have `max_worker_processes` set too low. The core of the problem here is
that one of the restarted background workers will cause master work to
start again (due to shared memory being reset, it'll get reinitialized)
and it'll start starting its own workers, while there may be too many
HTTP workers still running (before their silent termination when they
find out they are not "attached" to any master). To address this we
track that and wait until there are no more workers present during
master worker startup.