Problem: external dependencies
This makes initial builds much slower (CPM is cloning and is not
seemingly doing shallow cloning even when asked), and more difficult to
work on offline as well as packaging.

Solution: vendorize dependencies at specified versions

We don't have a lot of them anyway! Well, Python packages have some,
but a bearable amount.