Problem: omni_txn.retry doesn't set up the predicate infrastructure
This can be observed through the fact that we don't see `SIReadLock`s
being set by the transactions in `omni_txn.retry`.

Unless default isolation level is set as `serializable`.

This is pretty bad because effectively these transactions are not
participating in tracking of the reads across serializable transactions.

Solution: ensure we discard a potentially set snapshot
that was taken out in a different isolation mode.