Problem: consistency and performance of kube views
Currently, even though `omni_kube.api()` is `stable`, it does not
guarantee that it will be only called once per portal (~ statement)

This leads to potential inconsistencies when joining views (information
changes). Some of these are unavoidable if we call different API
endpoints, but it'd be really odd to get one when we're calling the same
API endpoint with the same parameters but something has changed already.

Besides, it slows it down as we make repeat requests

Solution: use omni_var's statement variables
to cache responses