Problem: omni_txn.retry may crash in some cases
In particular, when a prepared statement with omni_txn.retry was called,
after a few attempts, a crash was observed related to the query plan
being garbage:

```
0x000077f58ea84144 in extension_upgrade_hook (handle=0x7ffeaf343da0, pstmt=0x13cb7c0, queryString=<optimized out>, readOnlyTree=true, context=PROCESS_UTILITY_TOPLEVEL,
    params=0x13c7640, queryEnv=0x0, dest=0xb5c080 <donothingDR>, qc=0x7ffeaf344110) at /omnigres/extensions/omni/extension.c:132
132       switch (nodeTag(pstmt->utilityStmt)) {
```

In this particular case, the whole `pstmt` was observed to be garbage.

It was hard to reproduce this in automated testing.

Solution: don't restart transactions to get rid of wrong iso level

Instead, pop all the snapshots, and ensure the newly set snapshot will
account for an updated transaction isolation level.