Problem: running extension checks in CI
They are sometimes failing and this adds too much of maintaince burden.

We're almost ready to start publishing Omnigres (and other) packages
using [postgres.pm](https://github.com/postgres-pm/pgpm)

Scripts there already do the upgrade generation and this is the most
faulty process we currently have.

Solution: remove the extensions pipeline

This further forces us to make postgres.pm go live soon.

Question: how will this prevent us from not updating the version, for
which the existing pipeline had a check?

Answer: well, postgres.pm's script for Omnigres packages picks the
latest revision of the released version. Which means that it might
accidentally release the new version as the old one (but we'll observe
this through build increment) but then this also allows us to re-release
a version by publishing the latest version without updating.