Problem: running external Docker CI host
This is too much trouble:

* Sometimes it starts using too many resources and becomes unresponsive
* It requires connecting over Tailscale
* It imposes a very inconvenient PR approval workflow

Solution: run Docker tests on the same machine

For now, this means full test suite will only run on Linux. We will skip
Docker-related tests on macOS as it is currently not possible to run
Docker on it in the current CI setup we have.