Problem: occassional crash in `extension_upgrade_hook`
It appears that `pstmt` is freed when reaching the second invocaiton of
the hook (post standard process utility handling) in certain cases,
specifically in `CALL`/`DO` statement processing.

Solution: ensure we don't rely on `pstmt` being present
the second time around.

This should fix the problem, however, we should address it better in the
future as this is a very subtle problem others may run into it because
we offer `pstmt` at both ends of invocation and this is not safe.