Problem: independent extension building CI job
Sometimes it's failing to build.

Solution: ensure it gets the right cache from the get go

It's not perfect as we're still relying on something coming through the
cache (otherwise, we'll build it), but at least it should be an okay incrementally.
It'll not work well when minor versions of Postgres as misaligned.