Problem: omni_httpd doesn't work on multiple databases
If the extension is installed in more than one database, only one
database will have it running.

Solution: ensure omni_httpd tracking is database-scoped

This also transfered the atomic switchboard to omni_httpd itself (posing
a good question: should we remove this functionality from the `omni`
interface?)

Another concern here is that using multiple databases leads to a much
higher number of workers used. We use the same number of workers per
database. How should we handle this in the future, split the number
between databases? Introduce a total cap and a parameter for
per-database workers?

omni now zero-initializes shmem allocations. This is a safe change but
may potentially be a performance issue, though unlikely. If that will
ever be needed, we can introduce a flagged version of the call.