Problem: lack of HTTP server controls
Sometimes we need to shut HTTP server down (maintenance, security
issues, database administration tasks), but short of dropping the
extension and/or removing the listeners, this is not possible.

Removing the listeners is also insufficient for database administrative
tasks as the master worker will still be runing and preventing some of
the tasks from being accomplished.

Solution: introduce explicit stop/start procedures

This also necessitated a fix to `omni` that ensures that backend
initialization is folllowed by a commit and not a rollback to ensure
that the initialization sequence that latches onto "after commit" timing
will actually fire.