Problem: all schemas are open to PostgREST API
This is not how PostgREST operates. This was not really intentional but
rather a result of "let's get the basics done".

Solution: allow configuring schemas at the middleware

By default, no schemas are exposed, for safety reasons.

Other methods of schema switching were proposed, like supplying them in
the path. Definitely something to consider.

This change introduces `postgrest_settings` type which allows us to
configure things as needed. Expect it to grow over time.