Problem: does uuidv7 really belong in omni_id?
It's useful for identifiers, but can be used in many other scenarios.

Solution: move it omni_polyfill

We also add other functions from the family of functions.

IMPORTANT: omni_polyfill is now in its own schema and it needs to be
provisioned into the search path in a very specific (though
uncomplicated) way. See the documentation changes.