Problem: omnigres database in the container image
Omnigres containers are provisioned with the `omnigres` database.
Originally it was envisioned that the application will go into it,
but recent use shows that we use this one to bootstrap real application
database.

However, by default, we lack tools to do the provisioning in the
`omnigres` database.

Solution: install useful extensions for this

This should be sufficient to read configs, provision schemas and code
and do complex package installations.