Problem: accessing logs in container may be inconvenient
Sure, we can do `docker logs` but this requires docker itself and does
not lead to much automation.

Solution: collect logs into CSV and expose them through a table

This will stop printing log on the console, but one can exec `tail` on
the container to do something to the same effect.