Problem: omni_schema.assemble_schema progress is not observable
Since it is a PL/pgSQL function, it materializes the whole response and
we can't do value-per-call – and thus we can't provide just-in-time
reporting of steps taken.

This is not critical but it will affect the quality of life of external
tooling – we couldn't report what's going on until we're done.

Solution: emit notice messages

Closes #723