Problem: managing different services
Every service may have its own API when it comes to their operations.
This makes creation of generalized tooling difficult.

Solution: create a uniform service control plane

P.S. While at it, this is the first example of a template extension.
This is not necessarily because it really begs to be one, but to start
tacit transition to templated extensions to avoid problems normal
extensions have, allow multiple versions co-exist, etc.