Problem: omni_rest can be a bit slow
Not like "really slow" but I feel like it doesn't do enough.

Solution: don't construct query on every request

Provided that the request is the same, it doesn't make sense to do much
of this work in PL/pgSQL, reparsing the request and reconstructing the
query. It's wasteful.

For GET-only requests (for now), construct the query once and cache it.
In my limited local experimentation, this immediately 1.5x'd the throughput:
instead of 3.2K RPS I got about 5.1K.