Problem: managing API keys and other secrets
There's no standard facility provided by Omnigres, and this is very
common functionality.

Solution: add omni_credentials

It's a template extension that can be instantiated independently of the
extension itself (though by default, it instantiates itself)

It employs a neat trick useful in dev mode: you can instantiate a file
store that will be kept in sync with a file on the filesystem. In the
future, this can also potentially use omni_vfs (once
https://github.com/omnigres/omnigres/pull/710 is in).

This means that while developing, you can keep encrypted credentials file up to date.