Problem: importing a Python module after a failed attempt
If we try to load a module and it wasn't available, and then install it
and try again, omni_python might not see it.

Solution: invalidate loader caches

It is not very efficient to put this into every function call, but this
is currently handling the fact that installation may occur in a
different backend and therefore invalidating there wouldn't do anything.

While at it, let's make sure we don't add new `sys.path` if it is
already there.