Problem: stopped omni_httpd restarts by itself
When we create a new connection, the server starts back up

Solution: ensure we account for explicit control

The crux of the problem was that when we stop the server, it
flipped the switch of the started master worker back of zero.
However, new backends will not know the intention and will
eagerly start it again.

Now, we use this switch only for concurrency control as intended,
not as a marker.

While at this, an edge case of the same database name recreated
came up – we had allocated keys that depended on a name as opposed
to the OID, and they may have been not in the correct state. We now
rely on OID.