Problem: event messaging between different systems
Or, rather, the lack of universality. How do we know what kind of event is this
event?

Solition: add initial support for CloudEvents

It's simple and its getting some support and use throughout the industry.

At this point, this is a proof-of-concept that allows for basic publishing of events,
with publishers such as "NOTICE publisher" handling their delivery.