Problem: spontaneous omni_httpd misbehavior during stoppage
When stopping omni_httpd _very_ shortly after it has been started, we
have observed:

* Starts proceeding anyway
* Occassional segmentation faults.

Solution: ensure we guard for race conditions in this scenario better

* Replace atomic switchboard to locks for startup control

This is not in the hot path and there's no real atomicity to these sequences as
deciding on a master to start the background worker is disconnected from the
actual startup. This also doesn't preserve the intention – do we want to start the server?

* HTTP workers were getting blocked trying to receive FDs from the master worker

However, once ordered to shut down, master worker no longer serves these, but
we were incorrectly trying to wait for the response, without checking whether
the HTTP worker has been ordered to shut down.

* Shut down during HTTP worker startup

We've incorrectly handled the case where master worker is being ordered to shut
down while waiting for the initial response from HTTP workers. It proceeds to the
termination sequence, however, the `http_workers_started` flag is not set and it will
not attempt to shut any of them down, leading to orphaned workers that block the database
and do nothing.