Problem: omni_httpd worker sometimes segfaults during startup
When stopping omni_httpd _very_ shortly after startup due to dropping of the extension,
occassionally the worker will segfault.

Solution: ensure we handle errors during startup

What happened was that we automatically let Postgres handle exceptions arising
from the startup sequence in the worker, and it was calling `proc_exit` which was doing
shared memory reclamation. If SIGTERM arrived during that time, our signal handler would
get called and it tries to access that memory (since `BackendInitialized` was not cleared)

We're now handling this error ourselves and make sure to neutralize the SIGTERM handler
before we exit Postgres.

It might sound a bit concerning that `BackendInitialized` still gets set to `true` in this
scenario in the first place, but it is somewhat expectable, I think. We use some APIs that
are not involving the executor or related parts so `omni` does not get a chance to see that
`omni_httpd` is gone. Ideally, we should figure out a way to engage `omni` after we're set
in the snapshot.