Problem: deploying files from a filesystem in a different database
We have a case where we have to copy files from one filesystem to another
(table_fs) in a different database just so that database can use the files.

However, that puts a lot of burden onto those who implement it. See
`omni_schema.assemble_schema` for an example of that.

Solution: allow for remote file system access

As long as you can access a database over dblink, simply proxy calls to that
other database. Let's call it a "remote filesystem"

The design may not yet be ideal. I am in particular not 100% happy with the
idea of the "constructor" that we pass as a string to be able to indicate a
filesystem remotely. It's probably not the most terrible thing, but it irks me
as any text being passed to the other side. But that's dblink for you – it does
that anyway. The crux of this problem is that technically file systems are not
registered in the database (they _may be_, but that's an implementation
detail.)

There's a consideration to introduce "mountable" global filesystem that will somehow
glue all possible filesystems together, providing a global registration of all of them
at mountpoints. But we don't have it yet and we don't know if it will be the only way
to access mounted filesystems or not.

I had an idea that maybe we can create the type locally and then convert it
into a constructor, but that's complicated: the constructors are allowed to create database
objects or insert/update records. We might not want that done locally.

We may consider finding other ways to encode constructors that are better than
just passing a string to inject.

So let's just try to get going with the minimal possible thing.